Added components for vectorized dot product, cross product, and matrix vector product.
Added a new function to array_utils that tests for connection compatibility.  It does so by stripping aways 1s from the shape of the input/output that are adjacent
to the beginning/end of the shape tuple.  That is to say, a (1, 1, 3, 1, 3, 1, 1) can be connected to a (3, 1, 3) or a (1, 3, 1, 3, 1) since the underlying data is
stored in the same order.

Added docs for the new components.