Add back the "err_on_singular" option so that PETScDirectSolver is better compatible with DirectSolver methods and remove redundant method
Not all DirectSolvers will have the ability to just raise a warning and elegantly continue when there are NaNs or infs in the factorization or solve. So rather than removing the option from the PETScDirectSolver, I added it back in but made it so that it can only be True (with a descriptive error raise if it's not). Since the methods on the DirectSolver expect this option to exist, this approach makes the PETScDirectSolver more easily compatible with its parent's methods. Now we can directly use the "_inverse" method from the parent and reduce duplicate lines of code.