Add tests for the COBYQA solver option
Use the same tests as were used with COBYLA. However note that the one test that had no bounds for one of the design variables will not work with no bounds. The +/- 1e30 is too large for the COBYQA algorithm to work properly, so instead for that test we give it very large, but not as large bounds.