First pass at adding validate methods to the System class
Validate is meant to be used to check the values of any inputs / outputs in the model. Can be run after the run_model / run_driver to check if the inputs / outputs are "good". Running "run_validation" will run "validate" on each system all the way down the hierarchy from the system where it was called.