Update run_validation method
Update the validation method so that it captures all warnings and errors raised during calls to "validate" methods, then later provides a printout or raises an error showing all of the captured warnings / errors (if no errors were raised, then the warnings are just printed out). Also added a custom ValidationError for when it does raise an error. Added tests to go along with it