Removed test_lint_peps.py and test_lint_docstrings.py.
From now one, we will use ruff to check these in pre-commit and on CI.

The pre-commit tools can be installed in the pixi environment by running "pixi install-openmdao-dev" after installing the environment.

A series of ignores are included in pyproject.toml. We should work to reduce these over time, but for now they mostly replicate the behavior of the previous lint tests, except for a few minor differences.

Some functions in code_review, devtools, and docs were changed to satisfy linting.